Refactor MockResponse class into a new helpers module and update imports in test_client
